requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=0.1.0
openai>=1.30.0
httpx>=0.27.0
asyncpg>=0.29.0
redis[asyncio]>=5.0.0
python-telegram-bot>=21.0.0
pydub>=0.25.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
structlog>=24.0.0
tenacity>=8.2.0
python-multipart>=0.0.9
pgvector>=0.3.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
